Added GameRepository Game Integration tests
