+++ src/main/java/ch/uzh/ifi/hase/soprafs23/constant/GameStatus.java
@@ -1,5 +1,5 @@
 package ch.uzh.ifi.hase.soprafs23.constant;
 
 public enum GameStatus {
-  INPROGRESS, FINISHED
+  INLOBBY, INPROGRESS, FINISHED
 }
